Restructure application for deployment
