+++ requirements.txt
@@ -1,0 +1,7 @@
+certifi==2018.8.24
+click==6.7
+Flask==1.1.1
+itsdangerous==0.24
+Jinja2==2.10.3
+MarkupSafe==1.0
+Werkzeug==0.16.0
