+++ static/gamestatus.js
