+++ templates/layout.html
@@ -49,7 +49,7 @@
                     </ul>
                 {% endif %} -->
                     <ul class="navbar-nav mr-auto mt-2">
-                        <li class="nav-item"><a class="nav-link" href="/startgame">Select game</a></li>
+                        <li class="nav-item"><a class="nav-link" href="/selectgame">Select game</a></li>
                         <li class="nav-item"><a class="nav-link" href="/history">History</a></li>
                         <li class="nav-item"><a class="nav-link" href="/scoreboard">Scoreboard</a></li>
                     </ul>
